refactor: consolidate FLUX attention into single module with inheritance
Move FluxJointAttention from modules/attention.py into
models/flux/attention.py and refactor Flux2ParallelSelfAttention to
extend it instead of nn.Module.

Key changes:
- Flux2ParallelSelfAttention now inherits FluxJointAttention, reusing
  QK norm, RoPE, backend dispatch, and Ulysses parallelism (~80 lines removed)
- FluxJointAttention moved to flux/attention.py alongside its only consumers
- modules/attention.py retains base Attention class (shared with WAN)
- All imports updated across transformer_flux.py, transformer_flux2.py,
  __init__.py, and test_flux_attention.py

Verified on H200 and B200: FLUX.2 BF16, FLUX.2 FP8, FLUX.1 BF16 regression.
