+++ tensorrt_llm/_torch/visual_gen/models/flux/__init__.py
@@ -1,13 +1,14 @@
 # SPDX-FileCopyrightText: Copyright (c) 2022-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from .attention_flux2 import Flux2ParallelSelfAttention, Flux2PosEmbed
+from .attention import Flux2ParallelSelfAttention, Flux2PosEmbed, FluxJointAttention
 from .pipeline_flux import FluxPipeline
 from .pipeline_flux2 import Flux2Pipeline
 from .transformer_flux import FluxTransformer2DModel
 from .transformer_flux2 import Flux2Transformer2DModel
 
 __all__ = [
+    "FluxJointAttention",
     "FluxPipeline",
     "FluxTransformer2DModel",
     "Flux2Pipeline",
