+++ tensorrt_llm/_torch/visual_gen/modules/attention.py
@@ -1,5 +1,5 @@
 from enum import Enum
-from typing import TYPE_CHECKING, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Optional, Tuple
 
 import torch
 import torch.nn as nn
@@ -324,190 +324,3 @@ class Attention(nn.Module):
         out = self._attn_impl(q, k, v, batch_size, seq_len, kv_seq_len)
         out = self.to_out[0](out)
         return out
-
-
-class FluxJointAttention(Attention):
-    """Joint attention module for FLUX transformer models (FLUX.1 and FLUX.2).
-
-    Extends base Attention with:
-    - Text-stream QKV projection (add_qkv_proj) for dual-stream blocks
-    - FLUX-style RoPE on concatenated text+image tokens
-    - pre_only mode for single-stream blocks (no output projection)
-
-    For dual-stream blocks: Returns (img_attn_output, txt_attn_output)
-    For single-stream blocks: pre_only=True, returns attention output only
-    """
-
-    def __init__(
-        self,
-        hidden_size: int,
-        num_attention_heads: int,
-        head_dim: int = 128,
-        bias: bool = False,
-        added_kv_proj_dim: Optional[int] = None,
-        added_proj_bias: bool = True,
-        out_bias: bool = True,
-        eps: float = 1e-6,
-        pre_only: bool = False,
-        config: Optional["DiffusionModelConfig"] = None,
-        layer_idx: int = 0,
-    ):
-        super().__init__(
-            hidden_size=hidden_size,
-            num_attention_heads=num_attention_heads,
-            head_dim=head_dim,
-            qkv_mode=QKVMode.FUSE_QKV,
-            qk_norm=True,
-            qk_norm_mode="per_head",
-            eps=eps,
-            bias=bias,
-            out_bias=out_bias,
-            config=config,
-            layer_idx=layer_idx,
-        )
-
-        self.pre_only = pre_only
-        self.added_kv_proj_dim = added_kv_proj_dim
-
-        # Delete output projection for single-stream blocks
-        if self.pre_only:
-            del self.to_out
-
-        # Text-stream projections for joint attention (dual-stream blocks only)
-        if added_kv_proj_dim is not None:
-            self.add_qkv_proj = Linear(
-                added_kv_proj_dim,
-                3 * self.q_dim,
-                bias=added_proj_bias,
-                dtype=self.dtype,
-                quant_config=self.quant_config,
-                skip_create_weights_in_init=self.skip_create_weights_in_init,
-                force_dynamic_quantization=self.force_dynamic_quantization,
-                disable_deep_gemm=True,
-                weights_loading_config=WeightsLoadingConfig(
-                    weight_mode=WeightMode.FUSED_QKV_LINEAR
-                ),
-                fused_weight_shard_indices_mapping={
-                    "q": (0, self.q_dim),
-                    "k": (self.q_dim, self.q_dim),
-                    "v": (2 * self.q_dim, self.q_dim),
-                },
-            )
-
-            self.norm_added_q = RMSNorm(
-                hidden_size=head_dim, eps=eps, dtype=self.dtype, has_weights=True
-            )
-            self.norm_added_k = RMSNorm(
-                hidden_size=head_dim, eps=eps, dtype=self.dtype, has_weights=True
-            )
-
-            self.to_add_out = Linear(
-                self.q_dim,
-                added_kv_proj_dim,
-                bias=out_bias,
-                dtype=self.dtype,
-                quant_config=self.quant_config,
-                skip_create_weights_in_init=self.skip_create_weights_in_init,
-                force_dynamic_quantization=self.force_dynamic_quantization,
-                disable_deep_gemm=True,
-            )
-
-    @staticmethod
-    def _apply_rope(
-        x: torch.Tensor,
-        freqs_cos: torch.Tensor,
-        freqs_sin: torch.Tensor,
-    ) -> torch.Tensor:
-        """Apply FLUX-style rotary embeddings to a 4D tensor.
-
-        Args:
-            x: Input [B, S, H, D]
-            freqs_cos: Cosine frequencies [S, D]
-            freqs_sin: Sine frequencies [S, D]
-        """
-        x_fp32 = x.float()
-        cos = freqs_cos.float().unsqueeze(0).unsqueeze(2)  # [1, S, 1, D]
-        sin = freqs_sin.float().unsqueeze(0).unsqueeze(2)
-
-        # Rotate pairs: [x0, x1, x2, x3, ...] -> [-x1, x0, -x3, x2, ...]
-        x_rotated = torch.stack([-x_fp32[..., 1::2], x_fp32[..., 0::2]], dim=-1).flatten(-2)
-        return (x_fp32 * cos + x_rotated * sin).to(x.dtype)
-
-    def forward(
-        self,
-        hidden_states: torch.Tensor,
-        encoder_hidden_states: Optional[torch.Tensor] = None,
-        attention_mask: Optional[torch.Tensor] = None,
-        image_rotary_emb: Optional[Tuple[torch.Tensor, torch.Tensor]] = None,
-    ) -> Union[torch.Tensor, Tuple[torch.Tensor, torch.Tensor]]:
-        """Forward pass of joint attention.
-
-        Args:
-            hidden_states: Image tokens [batch, img_seq, dim]
-            encoder_hidden_states: Text tokens [batch, txt_seq, dim] (for dual-stream)
-            attention_mask: Optional attention mask (unused, for API compat)
-            image_rotary_emb: Tuple of (cos, sin) for RoPE
-
-        Returns:
-            For dual-stream: Tuple of (img_attn_output, txt_attn_output)
-            For single-stream: Attention output tensor
-        """
-        batch_size = hidden_states.shape[0]
-
-        # Image QKV via base (returns 3D), then reshape to 4D for per-head ops
-        query, key, value = self.get_qkv(hidden_states)
-        query = query.unflatten(-1, (self.num_attention_heads, self.head_dim))
-        key = key.unflatten(-1, (self.num_attention_heads, self.head_dim))
-        value = value.unflatten(-1, (self.num_attention_heads, self.head_dim))
-
-        # Per-head QK normalization via base (per_head mode operates on 4D)
-        query, key = self.apply_qk_norm(query, key)
-
-        # Text QKV for joint attention (dual-stream blocks)
-        if encoder_hidden_states is not None and self.added_kv_proj_dim is not None:
-            txt_seq_len = encoder_hidden_states.shape[1]
-
-            encoder_qkv = self.add_qkv_proj(encoder_hidden_states)
-            enc_q, enc_k, enc_v = encoder_qkv.chunk(3, dim=-1)
-            enc_q = enc_q.unflatten(-1, (self.num_attention_heads, self.head_dim))
-            enc_k = enc_k.unflatten(-1, (self.num_attention_heads, self.head_dim))
-            enc_v = enc_v.unflatten(-1, (self.num_attention_heads, self.head_dim))
-
-            enc_q = _per_head_norm(enc_q, self.norm_added_q)
-            enc_k = _per_head_norm(enc_k, self.norm_added_k)
-
-            # Concatenate text + image for joint attention
-            query = torch.cat([enc_q, query], dim=1)
-            key = torch.cat([enc_k, key], dim=1)
-            value = torch.cat([enc_v, value], dim=1)
-
-        # Apply RoPE
-        if image_rotary_emb is not None:
-            freqs_cos, freqs_sin = image_rotary_emb
-            query = self._apply_rope(query, freqs_cos, freqs_sin)
-            key = self._apply_rope(key, freqs_cos, freqs_sin)
-
-        # Flatten 4D→3D for base _attn_impl
-        seq_len = query.shape[1]
-        query = query.flatten(2)
-        key = key.flatten(2)
-        value = value.flatten(2)
-
-        hidden_states = self._attn_impl(query, key, value, batch_size, seq_len)
-        hidden_states = hidden_states.to(query.dtype)
-
-        # Split and project outputs
-        if encoder_hidden_states is not None and self.added_kv_proj_dim is not None:
-            encoder_hidden_states_out, hidden_states = hidden_states.split(
-                [txt_seq_len, hidden_states.shape[1] - txt_seq_len], dim=1
-            )
-
-            if not self.pre_only:
-                hidden_states = self.to_out[0](hidden_states)
-            encoder_hidden_states_out = self.to_add_out(encoder_hidden_states_out)
-
-            return hidden_states, encoder_hidden_states_out
-        else:
-            if not self.pre_only:
-                hidden_states = self.to_out[0](hidden_states)
-            return hidden_states
