+++ tests/unittest/_torch/visual_gen/test_flux_attention.py
@@ -42,7 +42,7 @@ class TestFluxAttentionBackend(unittest.TestCase):
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
     def test_vanilla_backend_sanity(self):
         """Test FLUX attention works with VANILLA backend."""
-        from tensorrt_llm._torch.visual_gen.modules.attention import FluxJointAttention
+        from tensorrt_llm._torch.visual_gen.models.flux.attention import FluxJointAttention
 
         batch_size = 2
         seq_len = 256
@@ -91,7 +91,7 @@ class TestFluxAttentionBackend(unittest.TestCase):
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
     def test_trtllm_backend_sanity(self):
         """Test FLUX attention works with TRTLLM backend."""
-        from tensorrt_llm._torch.visual_gen.modules.attention import FluxJointAttention
+        from tensorrt_llm._torch.visual_gen.models.flux.attention import FluxJointAttention
 
         batch_size = 2
         seq_len = 256
@@ -140,7 +140,7 @@ class TestFluxAttentionBackend(unittest.TestCase):
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
     def test_backend_equivalence(self):
         """Test VANILLA and TRTLLM backends produce similar outputs."""
-        from tensorrt_llm._torch.visual_gen.modules.attention import FluxJointAttention
+        from tensorrt_llm._torch.visual_gen.models.flux.attention import FluxJointAttention
 
         batch_size = 1
         seq_len = 128
@@ -262,7 +262,7 @@ class TestFlux2AttentionBackend(unittest.TestCase):
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
     def test_flux2_vanilla_backend_sanity(self):
         """Test FLUX.2 attention works with VANILLA backend."""
-        from tensorrt_llm._torch.visual_gen.modules.attention import FluxJointAttention
+        from tensorrt_llm._torch.visual_gen.models.flux.attention import FluxJointAttention
 
         batch_size = 2
         seq_len = 128
